bcp ajout

+++ src/main/java/fr/hyu/niflheimMMO/mobs/MonstersListeners.java
@@ -1,9 +1,11 @@
 package fr.hyu.niflheimMMO.mobs;
 
 import fr.hyu.niflheimMMO.experience.ExperienceGetter;
+import fr.hyu.niflheimMMO.statsApplicator.AgilityApplicator;
 import fr.hyu.niflheimMMO.statsApplicator.StrenghtApplicator;
 import fr.hyu.niflheimPermissions.player.PlayerRankProfile;
 import org.bukkit.Bukkit;
+import org.bukkit.ChatColor;
 import org.bukkit.entity.Entity;
 import org.bukkit.entity.EntityType;
 import org.bukkit.entity.LivingEntity;
@@ -14,8 +16,6 @@ import org.bukkit.event.entity.EntityDamageByEntityEvent;
 import org.bukkit.event.entity.EntityDamageEvent;
 import org.bukkit.event.entity.EntityDeathEvent;
 
-import static java.lang.String.valueOf;
-
 public class MonstersListeners implements Listener {
 
     @EventHandler
@@ -25,10 +25,13 @@ public class MonstersListeners implements Listener {
         if (event.getDamager() instanceof Player) {
             Player player = (Player) event.getDamager();
             if (PlayerRankProfile.hasPermission(player, "niflheim.devlog")) {
-                event.setDamage(0);
+                event.setDamage(0); //Cancel des dégats du jeu pour appliquer les dégats du plugin
                 if (entity instanceof LivingEntity) {
+
                     LivingEntity livingEntity = (LivingEntity) entity;
-                    if(StrenghtApplicator.damageOnMob(entity, player, player.getItemInUse())) { ExperienceGetter.setExperienceEarned(livingEntity, player); livingEntity.setHealth(0.0);}
+                    MonsterProfile monsterProfile = new MonsterProfile(livingEntity, entity.getLocation());
+                    if(AgilityApplicator.isDodgeMob( monsterProfile)){ player.sendMessage(ChatColor.GRAY + "Esquivé !"); event.setCancelled(true);return;}
+                    if(StrenghtApplicator.damageOnMob(monsterProfile, player)) { ExperienceGetter.setExperienceEarned(livingEntity, player); livingEntity.setHealth(0.0); MonsterDrop.DropLoot(entity);}
                 }
             }
         }
@@ -51,6 +54,9 @@ public class MonstersListeners implements Listener {
     public void onMonsterDeath(EntityDeathEvent event) {
         LivingEntity entity = event.getEntity();
         EntityType entityType = entity.getType();
+
+        if(entity.getLastDamageCause() == null) return;
+
         if (entity.getKiller() instanceof Player) { //double kill avec les stats de on normal faut retirer ça
             ExperienceGetter.setExperienceEarned((LivingEntity) entity, entity.getKiller());
         }
